refactor: api와 통신하는 repository Remote로 네이밍 수정
