+++ app/src/main/java/woowacourse/shopping/data/repository/remote/RemoteOrderRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.repository.real
+package woowacourse.shopping.data.repository.remote
 
 import woowacourse.shopping.data.remote.source.OrderDataSourceImpl
 import woowacourse.shopping.data.source.OrderDataSource
@@ -6,7 +6,7 @@ import woowacourse.shopping.domain.repository.OrderRepository
 import woowacourse.shopping.utils.exception.LatchUtils.executeWithLatch
 import woowacourse.shopping.utils.exception.NoSuchDataException
 
-class OrderRepositoryImpl(
+class RemoteOrderRepositoryImpl(
     private val orderDataSource: OrderDataSource = OrderDataSourceImpl(),
 ) : OrderRepository {
     override fun orderShoppingCart(ids: List<Int>) {
